% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_split.R
\name{block_split}
\alias{block_split}
\title{Block Split}
\usage{
block_split(training_data, training_labels, K, seed)
}
\arguments{
\item{training_data}{A tibble of the data to be split into blocks containing \emph{X} and \emph{Y} coordinates}

\item{training_labels}{A vector containing the labels for the data given}

\item{K}{The number of folds to split the data into}

\item{seed}{A random seed to permute the blocks}
}
\value{
A list containing the following:

training_data: The training data with the blocks assigned in a column

labels: a list of which block each label corresponds to

snum: The number of blocks the images have been split into
}
\description{
Block Split
}
