% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVmaster.R
\name{CVmaster}
\alias{CVmaster}
\title{CV Master File}
\usage{
CVmaster(
  training_data,
  training_labels,
  classifier,
  K = 5,
  loss = "misclassification",
  split = c("image", "block"),
  type = c("class", "prob", "response"),
  thresh = 0.5,
  formula = TRUE,
  seed = 521,
  ...
)
}
\arguments{
\item{training_data}{A tibble of the data to be split into blocks containing \emph{X} and \emph{Y} coordinates}

\item{training_labels}{A vector containing the labels for the data given}

\item{classifier}{The \strong{name} of a generic classifier}

\item{K}{The number of folds to split the data into}

\item{loss}{A loss function (only misclassification error is used)}

\item{split}{Which split function should be used to split the data}

\item{type}{What type should be used for the predictions (depends on classifier)}

\item{thresh}{For response/prob predictions, what should be the binary threshold}

\item{formula}{Should the classifier call a formula or separate X and y (needed for some classifiers)}

\item{seed}{A random seed for reproducibility}

\item{...}{Additional arguments passed forward into the classifier call}
}
\value{
A list containing the following:

CV_loss: A dataframe containing the loss for each fold, average CV loss, and test loss

roc_obj: A ROC object constructed from the test_set predictions

cv_preds: The predictions from each of the K CV folds

cv_response_preds: If type == "response", the probability from the CV predictions

test_preds: The predictions from the test_set model trained on all the training data

test_response_preds: If type == "response", the probability from the test predictions

test_labels: The true labels from the test set
}
\description{
CV Master File
}
