% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_split.R
\name{image_split}
\alias{image_split}
\title{Image Split}
\usage{
image_split(data, labels, K)
}
\arguments{
\item{data}{A tibble of data to be split into blocks with \emph{X} and \emph{Y} coordinates}

\item{labels}{A vector of labels corresponding to the data above}

\item{K}{An integer number of splits to be made}
}
\value{
A list containing the following:

training data: A dataframe containing all of the unlabeled training data

labels: A vector containing the block corresponding to each data point

train_labels: A vector containing the binary training data labels

val_data: A dataframe containing the validation dataset

val_labels: A vector containing the labels for the validation dataset

test_data: A dataframe containing the test dataset

test_labels: A vector containing the labels for the test dataset

snum: The number of blocks the images have been split into
}
\description{
Image Split
}
